Move `useSkuVariations` to `/components/ui/SkuSelector`
